[Sprint 1][Task 5] refactor: move tests from routes.rs to separate files
