+++ auth-service/tests/api/login.rs
@@ -1,0 +1,13 @@
+use crate::helpers::TestApp;
+
+#[tokio::test]
+async fn post_login_should_suceed() {
+    let app = TestApp::new().await;
+    let response = app.post_login().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+    // assert_eq!(
+    //     response.headers().get("content-type").unwrap(),
+    //     "application/json"
+    // );
+}
