+++ auth-service/tests/api/main.rs
@@ -1,2 +1,7 @@
 mod helpers;
-mod routers;
+mod login;
+mod logout;
+mod root;
+mod signup;
+mod verify_2fa;
+mod verify_token;
