+++ auth-service/tests/api/verify_2fa.rs
@@ -1,0 +1,12 @@
+use crate::helpers::TestApp;
+#[tokio::test]
+async fn post_verify2fa_should_suceed() {
+    let app = TestApp::new().await;
+    let response = app.post_verify2fa().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+    // assert_eq!(
+    //     response.headers().get("content-type").unwrap(),
+    //     "application/json"
+    // );
+}
